+++ src/ts/main.ts
@@ -71,6 +71,7 @@ import {StorageUtils} from './storageutils';
 import {ArrayUtils as Dummy} from './utils'; // tslint:disable-line
 import {SubtitleSettingsResetButton} from './components/subtitlesettings/subtitlesettingsresetbutton';
 import {ListBox} from './components/listbox';
+import {DemoFactory} from './demofactory';
 
 // Object.assign polyfill for ES5/IE9
 // https://developer.mozilla.org/de/docs/Web/JavaScript/Reference/Global_Objects/Object/assign
@@ -125,6 +126,7 @@ const playerui = {
   Component,
   Container,
   ControlBar,
+  DemoFactory,
   ErrorMessageOverlay,
   FullscreenToggleButton,
   HugePlaybackToggleButton,